requirements.txt
bs4
python-dotenv
langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-openai
langchain-groq
langchain-core
langchain-huggingface
langchain-chroma
faiss-cpu
fastapi
uvicorn
langserve
sse_starlette
transformers
streamlit
pypdf
arxiv
wikipedia